backend/requirements.txt
asgiref==3.10.0
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
Django==5.2.8
djangorestframework==3.16.1
filelock==3.19.1
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
psycopg2-binary==2.9.11
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
